started to edit python code to communicate with arduino and wrote a new arduino file to communicate with MPU-6050 as well as snake.py

+++ serial-wr/serial-rw/serial-rw.ino
@@ -10,7 +10,7 @@
 
 // LED status
 int led_status = HIGH;
-int incomingByte = 0;
+byte incomingByte = 0;
 
 // the setup routine runs once when you press reset:
 void setup() {
